feat: Add CBS module
Bit of a cirularity issue. JaBbA needs purity/ploidy estimate via ASCAT
but ASCAT needs CBS coverages for tumor-normal pair from dryclean.

Did not implement the merging of the dryclean output channels for
tumor-normal. This code assumes there is a channel which merges the two
(with meta) into a tuple