fix: Fixes conditional switches for JaBbA
JaBbA CLI has a bunch of conditional switches which needed to be
implemented as strings in the process script.

Also changes blacklist_junctions parameter to a val so that a default
value of "NULL" can be used (instead of "/dev/null" which results in nextflow
binding the local "/dev" directory (which may have limited permissions if on an hpc)
to the container and overwriting its native "/dev/null". This in turn results in a bug
where /dev/null is inaccessible.